[chore] Move query service to higher location
