+++ cmd/jaeger/internal/extension/jaegerquery/internal/adjuster/hash.go
