+++ cmd/jaeger/internal/extension/jaegerquery/internal/adjuster/hash_test.go
