+++ cmd/jaeger/internal/extension/jaegerquery/internal/adjuster/libraryattributes.go
